# Use IdentityHashMap for grouping.