Resources are essentially Singletons, and TracerProvider ensures the same for InstrumentationLibraryInfo. We may as well use an IdentityHashMap when grouping them to not rely on performance characteristics of `hashCode`.