@jkwatson 2021-08-13
> Just random tidbit, first time reading through the code for `IdentityHashMap` - I didn't know JDK provides a `HashMap` that uses open addressing! The javadocs mention "For many JRE implementations and operation mixes, this class will yield better performance than HashMap (which uses chaining rather than linear-probing)."
> 
> So figured worth running the benchmarks, looks about the same
> 
> After
> 
> ```
> Benchmark                                                  Mode  Cnt     Score     Error   Units
> SpanAdapterBenchmark.toProto                               avgt   45     2.536 ±   0.071   us/op
> SpanAdapterBenchmark.toProto:·gc.alloc.rate                avgt   45  1472.886 ±  39.027  MB/sec
> SpanAdapterBenchmark.toProto:·gc.alloc.rate.norm           avgt   45  5869.336 ±  16.036    B/op
> SpanAdapterBenchmark.toProto:·gc.churn.G1_Eden_Space       avgt   45  1473.465 ±  51.696  MB/sec
> SpanAdapterBenchmark.toProto:·gc.churn.G1_Eden_Space.norm  avgt   45  5870.932 ± 127.178    B/op
> SpanAdapterBenchmark.toProto:·gc.churn.G1_Old_Gen          avgt   45     0.003 ±   0.001  MB/sec
> SpanAdapterBenchmark.toProto:·gc.churn.G1_Old_Gen.norm     avgt   45     0.013 ±   0.004    B/op
> SpanAdapterBenchmark.toProto:·gc.count                     avgt   45   452.000            counts
> SpanAdapterBenchmark.toProto:·gc.time                      avgt   45   203.000                ms
> ```
> 
> Before
> 
> ```
> Benchmark                                                  Mode  Cnt     Score     Error   Units
> SpanAdapterBenchmark.toProto                               avgt   45     2.557 ±   0.028   us/op
> SpanAdapterBenchmark.toProto:·gc.alloc.rate                avgt   45  1345.528 ±  14.370  MB/sec
> SpanAdapterBenchmark.toProto:·gc.alloc.rate.norm           avgt   45  5416.003 ±   0.004    B/op
> SpanAdapterBenchmark.toProto:·gc.churn.G1_Eden_Space       avgt   45  1346.696 ±  32.660  MB/sec
> SpanAdapterBenchmark.toProto:·gc.churn.G1_Eden_Space.norm  avgt   45  5420.231 ± 110.335    B/op
> SpanAdapterBenchmark.toProto:·gc.churn.G1_Old_Gen          avgt   45     0.003 ±   0.001  MB/sec
> SpanAdapterBenchmark.toProto:·gc.churn.G1_Old_Gen.norm     avgt   45     0.013 ±   0.003    B/op
> SpanAdapterBenchmark.toProto:·gc.count                     avgt   45   390.000            counts
> SpanAdapterBenchmark.toProto:·gc.time                      avgt   45   178.000                ms
> ```

Allocations look worse with the IDH, though. That's surprising to me!