@anuraaga 2021-08-14
I think it's because the IDH defaults to a bit bigger backing array. It doesn't use Node wrappers though so it's some more bytes but less objects overall. I could try tweaking the size but the alloc change didn't seem big enough to try that to me.