@jkwatson 2021-08-14
I wonder if we could optimize to a very time initial size. 99.99% of the time there will only be one entry in the map.