@anuraaga 2021-08-14
For now I've matched the capacity of HashMap. I agree in practice we only have 1 resource and <5 instrumentation libraries, so we could possibly optimize more but will explore in another PR.

Alloc rate is now similar (less corresponds to the lack of Node objects)

```
Benchmark                                                  Mode  Cnt     Score     Error   Units
SpanAdapterBenchmark.toProto                               avgt   45     2.554 ±   0.126   us/op
SpanAdapterBenchmark.toProto:·gc.alloc.rate                avgt   45  1353.807 ±  66.727  MB/sec
SpanAdapterBenchmark.toProto:·gc.alloc.rate.norm           avgt   45  5400.003 ±   0.004    B/op
SpanAdapterBenchmark.toProto:·gc.churn.G1_Eden_Space       avgt   45  1355.947 ±  77.782  MB/sec
SpanAdapterBenchmark.toProto:·gc.churn.G1_Eden_Space.norm  avgt   45  5408.781 ± 159.303    B/op
SpanAdapterBenchmark.toProto:·gc.churn.G1_Old_Gen          avgt   45     0.003 ±   0.001  MB/sec
SpanAdapterBenchmark.toProto:·gc.churn.G1_Old_Gen.norm     avgt   45     0.010 ±   0.003    B/op
SpanAdapterBenchmark.toProto:·gc.count                     avgt   45   347.000            counts
SpanAdapterBenchmark.toProto:·gc.time                      avgt   45   170.000                ms
```